requirements.txt
discord.py
colorama
pillow
python-dotenv
requests
aiosqlite
PyNaCl
yt-dlp 
spotipy
wavelink
beautifulsoup4
aiohttp
qrcode[pil]
google-generativeai
gTTS
imageio-ffmpeg
Flask